[11382] 꼬마 정민

+++ Baekjoon/11382/11382.cpp
@@ -1,13 +1,13 @@
 #include <iostream>
 
 int main() {
-  long a{};
-  long b{};
-  long c{};
+  long long a{};
+  long long b{};
+  long long c{};
 
   std::cin >> a >> b >> c;
 
-  int sum = a + b + c;
+  long long sum = a + b + c;
 
   std::cout << sum << std::endl;
